fix: keep pitch in float32 so CUDA training works
librosa.pyin returns float64, so target_pitch promoted the pitch loss - and
through it the whole generator loss - to double. CPU autograd tolerated the
mismatch, but CUDA rejects double gradients flowing into float parameters:

    RuntimeError: Found dtype Double but expected Float

save_features now writes pitch as float32, and VoiceDataset casts on load so
shards written before this change keep working without re-preprocessing.

+++ src/voice_swap/preprocess.py
@@ -244,9 +244,11 @@ def _save_feature_shard(
 
 def save_features(features: list[dict], path: Path) -> None:
     """Save extracted features to disk."""
+    # librosa.pyin returns float64; store float32 so the pitch loss does not
+    # promote the whole generator loss to double (CUDA rejects double grads).
     batch = {
-        "mel": torch.tensor(np.array([f["mel"] for f in features])),
-        "pitch": torch.tensor(np.array([f["pitch"] for f in features])),
+        "mel": torch.tensor(np.array([f["mel"] for f in features]), dtype=torch.float32),
+        "pitch": torch.tensor(np.array([f["pitch"] for f in features]), dtype=torch.float32),
     }
     torch.save(batch, path)
